Add UrlStorage
This class is responsible for writing and reading the stored urls.
It stores them in memory using two hashes.

The first hash (@store) uses the `url` as the key while the second one
(@store_reversed) uses the `short_url` as the key, the reason behind is
that it will be faster to access the value of a given `short_url` as the
complexity will be of O(1) instead of O(n) (if using a single hash).